Optimize for size and hack-friendliness.
The Y value does not need to be matched for the number of things being
refilled anymore.  It's initialized to 0, any item that executes ADC does INY,
and a non-zero value after all items triggers "run next frame" logic.

As such, we now see "was full" instead of "is now full" as the status, so we
don't have to store the extra flag in `$05F9`, saving a few bytes.  `inc_item`
didn't even expand to give us this more-convenient info.